patched mrjob.zip into py_files. moved self._setup and related to mrjob/bin.py
this abuses _parse_setup_and_py_files(), which is really just for creating
self._setup at init time. we can do the PYTHONPATH stuff in the setup wrapper
rather than at init time
